Restructuring.
